Default Heltec T114 to Fixed Pin
Heltec T114 has a no-screen option available for purchase. At
present the Bluetooth PIN on such a device is set to random, which
makes it challenging for users to connect.

We currently don't have code to detect (lack of) a SPI-based screen.
For now, set T114 to use a fixed pin, which users can later change.

Fixes https://github.com/meshtastic/firmware/issues/4565